+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.14.4
 
 Changes:
   - Add additional output columns to gap-junction runs. See FUNCZ-211_.
+  - Fix executions for circuits with only one synapse class. See FUNCZ-218_.
 
 Version 0.14.3
 ==============
@@ -168,3 +169,4 @@ First working version with 3 base filters:
 .. _FUNCZ-201: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-201
 .. _FUNCZ-209: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-209
 .. _FUNCZ-211: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-211
+.. _FUNCZ-218: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-218